Replaced use of awk in pacman-list-pkgs alias
Previously the formatting of the output of pacman-list-pkgs wouldn't
be correct (without the version numbers) as a result of the '$1' in
the call to awk being unescaped. Here, we dispense with awk altogether
by using pacman's -Qq option to remove the version numbers from the
output.

Also removed usage of -Qt as it would result in the alias not listing
all of the packages it should have been.