Heavily refactor and simplify bash configuration
As we're aiming to permanently switch to zsh, we only really need
the bare minimum of implementation features in our bash
configuration to do work when on machines without zsh.

Unfortunately this is made a lot more complex by the need to hack
around applications refusing to support XDG path conventions (one of
which being bash itself).