Remove agent support
Want to use Funtoo keychain for both ssh and gpg agent management, but
seems to interact weirdly with newer versions of gpg and the built-in
system service for gpg-agents under some distros. Recommend implementing
this in a platform specific way (for example, inside NixOS configuration
via services, or inside a `custom.login` file) in order to allow for
customising to a specific machine. I expect a similar approach with
regards to `gpg-agent.conf` and `ssh/config` also; this way you are
forced to think about your security carefully when configuring your
system.