Update README to be more complete & cohesive
README was a bit of a mess prior to this commit. To fix this we rewrote
large portions of the document, and added many sections explaining the
project at various levels of abstraction, with improved structuring.

A typo in the `tmux.conf` top-level comment about the custom
keybindings, which made it unclear how `<Prefix> <C-f>` compared to the
built-in `<Prefix> f` binding, has been fixed.

Additionally, URxvt and XTerm must always be installed together due to a
limitation of Xresources' default preprocessor causing errors when
including files that don't exist on the system (i.e. `.Xresources` as
provided in the theme expects both URxvt and XTerm `*.xrc` files to
always be present). This has been resolved by placing both
configurations in the same logical package `xapplications`.