Move `filetype indent plugin on` after plugin load
vim-plug handles turning this on anyway, but just in case we ever decide
to change plugin manager, we handle it sanely ourselves. Both syntax and
filetype indent plugin should be set to off at the start of our
configuration, and then turned on after all plugins are loaded in order
to guarantee everything gets loaded as expected.