Fix broken Vim which-key bindings
vim-which-key chokes on mappings that are more complex, so we fix this
by extracting these into their own functions, and then calling them from
the mappings.

Some prompts have been improved to offer completion and defaults, and a
new binding for yanking file name and not just file path has been added.