Fix embark and consult integration
`embark-consult` wasn't being loaded due to the presence of the hook, causing
embark to not show appropriate actions for candidates in the output of commands
such as `consult-buffer`. We choose to demand load it to avoid this issue, as
there doesn't appear to be a good way of hooking into embark being loaded.