Fix Emacs daemon with tree-sitter
Tree-sitter being enabled on all `prog-mode` derived major modes would cause
daemonized Emacs to hang on startup and timeout due to `tree-sitter-mode`
erroring as a result of there being a missing emacs-lisp grammar. We work around
this by demoting errors on the `prog-mode` hook to warning messages
instead (they are harmless in practice).