Remove reference to NixOS from the readme
Plans for this have been scuppered due to concerns I have over the
velocity of NixOS (particularly around how slow adoption of something as
critical as flakes has been - as well as how poor the documentation and
UX around flakes and modules are). Additionally I have concerns about
the security implications of the store for passwords and so on.