Fix `vim-plug` lockfiles blocking message
`plug#begin` needs to be called before loading the lock files, as lock
files use plug commands. Additionally `PlugUpdate!` as generated in the
lock files will cause a hanging message on starting up Vim unless it is
manually silenced, so we do so.

Also update plugin versions in the process.