Add simple Notmuch support to Emacs
This is not currently used because synchronizing everything across devices is
painful, but some useful functionality for switching between profiles was
written while I experimented with it, so I'm adding it anyway. I may continue
migrating to Notmuch for desktop/laptop email at some point, in which case this
will remain useful.