Remove libegit2 support from Emacs Magit
The author has currently removed it from Melpa due to it being experimental and
having some issues. We were also experiencing some issues revolving around the
shared library compilation process and paths. For now we have decided it isn't
really worth the complexity for a small performance boost. Perhaps we can re-add
this support later on.