Convert to zero-plugin minimal Vim configuration
We have long since switched to Emacs as a primary editor. It serves no real
purpose to heavily modify and extend the functionality of Vim when we intend to
only use it for small editing tasks. Advanced editing functionality should
instead be built into our Emacs configuration. Instead we prefer to remove all
plugins and only make slight alterations to the base configuration. The new
priorities of our Vim configuration are speed & minimalism.